requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
playwright==1.41.1
beautifulsoup4==4.12.2
lxml==5.1.0
requests==2.31.0
python-dotenv==1.0.0
stripe==7.11.0
gunicorn==21.2.0
pytest==7.4.3
black==23.12.1
flake8==7.0.0 